[CSL-531] Build uninstaller before installer for code signing
A previous iteration of this resulted in an uninstall.exe which worked as
expected, both via Windows' Uninstall Programs menu and as a standalone
executable (found in %TEMP%).

Additionally this includes an attempted yet untested fix for running the
uninstaller installer (tempinstaller.exe) in headless mode for CI purposes. The
obvious approach with Turtle failed for some reason (doesn't seem to be passing
"/S" argument as expected).