[DEVOPS-102] Run cardano-launcher.exe at highestAvailable level
Necessary for getting an update Installer.exe to elevate to admin
execution level when run by an admin user.

DEVOPS-118 tried to address this by setting requestExecutionLevel
to admin, but it doesn't take take effect when executed by a
non-elevated exe (in this case cardano-launcher.exe).

This approach has been verified on a Windows 10 Pro VM.