[CSL-2435] Use 'TextEncoder' to measure correct length of UTF-8 strings
JavaScript length method (with ES6) actually counts unicode punycodes of
the utf-16 encoding of the given string. Apart from messing up with
surrogate pairs (counting for instance 💩 as two characters because
encoded with a pair or punycodes), it also leads to unexpected results
when characters have a different UTF-8 and UTF-16 encoding. Since
requests are made using UTF-8 encoding, the Content-Length of these
requests should match the actual length when payloads are UTF-8 encoded.