[DDW-426] Estimate row heights in render method
The bug (expanded txs with wrong height in the virtual list) might have been caused by estimating the rows too late (in componentDidUpdate) when the virtual list already was setup. This change moves the estimation logic into the render before the virtual list is created. Hopefully this will resolve this issue