[DDW-426] Wrong transaction height when filtering (#2249)
* [DDW-426] Estimate row heights in render method

The bug (expanded txs with wrong height in the virtual list) might have been caused by estimating the rows too late (in componentDidUpdate) when the virtual list already was setup. This change moves the estimation logic into the render before the virtual list is created. Hopefully this will resolve this issue

* [DDW-426] Add changelog entry

Co-authored-by: Nikola Glumac <niglumac@gmail.com>