Fix disk space report for some Linux distros
The Linux installer runs Daedalus inside a chroot, which lives in a
temporary directory.

Some - perhaps most - Linux distros have the temporary directory as a
tmpfs (RAM drive).

So when Daedalus checks the free space in /, it is reporting the
amount of free space on the tmpfs - typically only a few GB.

This change makes Daedalus check the free space for the state
directory.

In any case, this is a good thing to do, because the state
directory (in the user's home directory) may be on a different
partition to /.