Suggestion to set -F1.25
The -F RTS param is used to indicate the next old generation heap size
after a collect. By default the parameter is 2. This means, when your
heap is 2 GB after old generation collect, heap will be sized to 4 GB
next time. This might be excessive for small memory devices. -F1.25
means the next heap will scale to 125% of the old generation live
memory.