Remove device param from TrezorConnect calls
When the param is specified, Trezor asks for passphrase on every call even if those happen in single session. This introduces room for error as user can provide different passphrase for exporting public key, and then different passphrase for signing transaction. Without this parameter (when path is managed by TrezorConnect) everything works as expected.